+++ NEWS.md
@@ -1,4 +1,10 @@
-# nychex 0.0.0.9000
+# nychex 0.2.0
+
+* Added `nyc_ct20_hex_sf`: hexagonal tile map of 2,271 census tracts (2020)
+  for all five NYC boroughs. Brooklyn and Queens are tiled as separate
+  boroughs.
+
+# nychex 0.1.0
 
 * Added `nyc_nta20_hex_sf`: hexagonal tile map of 262 NTA 2020 boundaries
   for all five NYC boroughs.